voice_privacy_eval/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
csvw==3.5.1
cycler==0.12.1
Cython==3.0.12
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
dlinfo==2.0.0
dtw-python==1.5.3
et_xmlfile==2.0.0
fastdtw==0.3.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
g2p-en==2.1.0
huggingface-hub==0.30.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
inflect==7.5.0
isodate==0.7.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
language-tags==1.2.0
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.21.1
onnxruntime-gpu==1.21.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pesq==0.0.4
phonemizer==3.3.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pycparser==2.22
pyparsing==3.2.3
pysptk==1.0.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-espeak==0.6.3
pytorch-lightning==2.5.1
pytz==2025.2
pyvad==0.2.0
pyworld==0.3.5
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rfc3986==1.5.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segments==2.3.0
sentencepiece==0.2.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
speechmos @ git+https://github.com/tarepan/SpeechMOS@ed25eacbfa42b99156c36ebec67a733b5dbb9b79
speechpy==2.4
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typeguard==4.4.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
utmosv2 @ git+https://github.com/sarulab-speech/UTMOSv2.git@e4527dfe654a8ea79cb3232e172f483a5eea5731
webrtcvad==2.0.10
xxhash==3.5.0
yarl==1.19.0
